+++ src/dfxvm/cli.rs
@@ -42,9 +42,17 @@ pub struct DefaultOpts {
     version: Option<Version>,
 }
 
-/// List installed versions of dfx
+/// List installed or available versions of dfx
 #[derive(Parser)]
-pub struct ListOpts {}
+pub struct ListOpts {
+    /// List the available versions.
+    #[arg(long)]
+    pub available: bool,
+
+    /// The maximum number of available versions to list, in reverse chronological order.
+    #[arg(long, default_value_t = 10, requires("available"))]
+    pub limit: usize,
+}
 
 /// Uninstall a version of dfx
 #[derive(Parser)]
@@ -89,7 +97,7 @@ pub async fn main(args: &[OsString], locations: &Locations) -> Result<ExitCode, 
     match cli.command {
         Command::Default(opts) => default(opts.version, locations).await?,
         Command::Install(opts) => install(opts.version, locations).await?,
-        Command::List(_opts) => list(locations)?,
+        Command::List(opts) => list(opts, locations).await?,
         Command::SelfCmd(opts) => match opts.command {
             SelfCommand::Update(_opts) => self_update(locations).await?,
             SelfCommand::Uninstall(opts) => self_uninstall(opts.yes, locations)?,