+++ src/dfxvm/manifest.rs
@@ -1,0 +1,13 @@
+use semver::Version;
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct Tags {
+    pub latest: Version,
+}
+
+#[derive(Deserialize)]
+pub struct Manifest {
+    pub tags: Tags,
+    pub versions: Vec<Version>,
+}