+++ src/dfxvm/update.rs
@@ -1,11 +1,10 @@
 use crate::dfxvm::default::set_default;
+use crate::dfxvm::manifest::Manifest;
 use crate::error::dfxvm::UpdateError;
 use crate::json::fetch_json;
 use crate::locations::Locations;
 use crate::settings::Settings;
 use reqwest::Url;
-use semver::Version;
-use serde::Deserialize;
 
 pub async fn update(locations: &Locations) -> Result<(), UpdateError> {
     let settings = Settings::load_or_default(&locations.settings_path())?;
@@ -20,14 +19,4 @@ pub async fn update(locations: &Locations) -> Result<(), UpdateError> {
     set_default(&latest_version, locations).await?;
 
     Ok(())
-}
-
-#[derive(Deserialize)]
-struct Tags {
-    latest: Version,
-}
-
-#[derive(Deserialize)]
-struct Manifest {
-    tags: Tags,
 }